backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
psycopg2-binary==2.9.1
python-dotenv==0.19.0
Werkzeug==2.0.1
PyJWT==2.3.0
Flask-Migrate
SQLAlchemy==1.4.23